+++ src/datadoc/backend/src/config.py
@@ -1,0 +1,1 @@
+"""Configuration management for Datadoc package."""
