+++ src/datadoc/backend/src/statistic_subject_mapping.py
@@ -10,7 +10,7 @@ from bs4 import BeautifulSoup
 from bs4 import ResultSet
 
 from datadoc.backend.external_sources.external_sources import GetExternalSource
-from datadoc.enums import SupportedLanguages
+from datadoc.backend.src.enums import SupportedLanguages
 
 if TYPE_CHECKING:
     import concurrent
