+++ src/datadoc/config.py
@@ -11,9 +11,9 @@ from typing import Literal
 from dotenv import dotenv_values
 from dotenv import load_dotenv
 
+from datadoc.backend.src.enums import DaplaRegion
+from datadoc.backend.src.enums import DaplaService
 from datadoc.constants import DAPLA_MANUAL_TEXT
-from datadoc.enums import DaplaRegion
-from datadoc.enums import DaplaService
 from datadoc.frontend.components.builders import build_link_object
 
 logging.basicConfig(level=logging.DEBUG, force=True)
