+++ src/datadoc/frontend/callbacks/utils.py
@@ -12,8 +12,8 @@ import arrow
 from datadoc_model import model
 
 from datadoc import config
-from datadoc import enums
 from datadoc import state
+from datadoc.backend.src import enums
 from datadoc.constants import DATE_VALIDATION_MESSAGE
 from datadoc.constants import VALIDATION_ERROR
 from datadoc.frontend.components.dataset_tab import build_dataset_tab
