+++ src/datadoc/frontend/fields/display_variables.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 import functools
 from enum import Enum
 
-from datadoc import enums
 from datadoc import state
+from datadoc.backend.src import enums
 from datadoc.frontend.fields.display_base import VARIABLES_METADATA_DATE_INPUT
 from datadoc.frontend.fields.display_base import VARIABLES_METADATA_MULTILANGUAGE_INPUT
 from datadoc.frontend.fields.display_base import FieldTypes
