+++ tests/backend/test_user_info.py
@@ -5,6 +5,8 @@ import pytest
 from faker import Faker
 
 from datadoc.backend.src import user_info
+from datadoc.backend.src.enums import DaplaRegion
+from datadoc.backend.src.enums import DaplaService
 from datadoc.backend.src.user_info import PLACEHOLDER_EMAIL_ADDRESS
 from datadoc.backend.src.user_info import DaplaLabUserInfo
 from datadoc.backend.src.user_info import JupyterHubUserInfo
@@ -13,8 +15,6 @@ from datadoc.backend.src.user_info import UserInfo
 from datadoc.config import DAPLA_REGION
 from datadoc.config import DAPLA_SERVICE
 from datadoc.config import JUPYTERHUB_USER
-from datadoc.enums import DaplaRegion
-from datadoc.enums import DaplaService
 
 
 @pytest.fixture()
