+++ tests/frontend/callbacks/test_callbacks_utils.py
@@ -4,7 +4,7 @@ import pytest
 from dash import html
 from datadoc_model import model
 
-from datadoc.enums import LanguageStringsEnum
+from datadoc.backend.src.enums import LanguageStringsEnum
 from datadoc.frontend.callbacks.utils import find_existing_language_string
 from datadoc.frontend.callbacks.utils import get_language_strings_enum
 from datadoc.frontend.callbacks.utils import render_tabs
