+++ tests/frontend/callbacks/test_variables_callbacks.py
@@ -14,8 +14,8 @@ from datadoc_model import model
 from datadoc_model.model import LanguageStringTypeItem
 from pydantic_core import Url
 
-from datadoc import enums
 from datadoc import state
+from datadoc.backend.src import enums
 from datadoc.backend.src.model_validation import ObligatoryVariableWarning
 from datadoc.frontend.callbacks.variables import accept_variable_metadata_date_input
 from datadoc.frontend.callbacks.variables import accept_variable_metadata_input
